refactors project and extracts classes out to their own files. adds gophish api functionality to import users. updates requirements.txt
